+++ R/Collect.reddit.R
@@ -102,9 +102,10 @@ Collect.reddit <-
     }
 
     class(threads_df) <- append(c("datasource", "reddit"), class(threads_df))
-    if (writeToFile) {
-      write_output_file(threads_df, "rds", "RedditData", verbose = verbose)
-    }
+    
+    meta_log <- c(collect_log, paste0(format(Sys.time(), "%a %b %d %X %Y")))
+    
+    if (writeToFile) write_output_file(threads_df, "rds", "RedditData", verbose = verbose, log = meta_log)
 
     msg("Done.\n")